// src/interfaces/builder-search.interface.ts
export interface BuilderSearch {
    id: string;
    builderId: string;
    builderName: string;
    projectId: string;
    projectName: string;
    projectRera: string;
    brokerId: string;
    brokerName: string;
    brokerRera: string;
    _createdDate?: string;
    _updatedDate?: string;
  }
  